refactor(adapters): extract data helpers to helpers.rs [#305-4/7]
- Move count_unique_packages and find_upgrade_action as pub(super) free functions
- Move test_count_unique_packages to helpers.rs
- Update mod.rs to call helpers:: functions
- Remove unused IntroducedByView and UpgradeEntryView from top-level imports

Closes #326
